Backend/requirements.txt
APScheduler==3.11.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coverage==7.7.1
dotenv==0.9.9
exceptiongroup==1.2.2
Faker==37.0.2
Flask==3.1.0
Flask-APScheduler==1.13.1
Flask-Cors==4.0.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
onesignal-python-api==2.0.2
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.3.5
pytest-cov==6.0.0
pytest-mock==3.14.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
responses==0.25.7
six==1.17.0
tomli==2.2.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
Werkzeug==3.1.3
